+++ json_serializable/test/default_value/implicit_default_value.g.dart
@@ -89,7 +89,7 @@ Map<String, dynamic> _$DefaultValueImplicitToJson(
   'fieldMapSimple': instance.fieldMapSimple,
   'fieldMapListString': instance.fieldMapListString,
   'durationField': instance.durationField.inMicroseconds,
-  'fieldEnum': _$GreekEnumMap[instance.fieldEnum]!,
+  'fieldEnum': $enumEncode(_$GreekEnumMap, instance.fieldEnum),
   'constClass': instance.constClass,
   'valueFromConverter': const ConstClassConverter().toJson(
     instance.valueFromConverter,