+++ json_serializable/test/supported_types/input.type_enumtype.g.dart
@@ -15,8 +15,8 @@ SimpleClass _$SimpleClassFromJson(Map<String, dynamic> json) => SimpleClass(
 
 Map<String, dynamic> _$SimpleClassToJson(SimpleClass instance) =>
     <String, dynamic>{
-      'value': _$EnumTypeEnumMap[instance.value]!,
-      'withDefault': _$EnumTypeEnumMap[instance.withDefault]!,
+      'value': $enumEncode(_$EnumTypeEnumMap, instance.value),
+      'withDefault': $enumEncode(_$EnumTypeEnumMap, instance.withDefault),
     };
 
 const _$EnumTypeEnumMap = {