+++ json_serializable/test/supported_types/input.type_list.g.dart
@@ -310,7 +310,9 @@ SimpleClassOfEnumType _$SimpleClassOfEnumTypeFromJson(
 Map<String, dynamic> _$SimpleClassOfEnumTypeToJson(
   SimpleClassOfEnumType instance,
 ) => <String, dynamic>{
-  'value': instance.value.map((e) => _$EnumTypeEnumMap[e]!).toList(),
+  'value': instance.value
+      .map((e) => $enumEncode(_$EnumTypeEnumMap, e))
+      .toList(),
 };
 
 const _$EnumTypeEnumMap = {
@@ -331,7 +333,9 @@ SimpleClassNullableOfEnumType _$SimpleClassNullableOfEnumTypeFromJson(
 Map<String, dynamic> _$SimpleClassNullableOfEnumTypeToJson(
   SimpleClassNullableOfEnumType instance,
 ) => <String, dynamic>{
-  'value': instance.value?.map((e) => _$EnumTypeEnumMap[e]!).toList(),
+  'value': instance.value
+      ?.map((e) => $enumEncode(_$EnumTypeEnumMap, e))
+      .toList(),
 };
 
 SimpleClassOfEnumTypeNullable _$SimpleClassOfEnumTypeNullableFromJson(